### 2022-11-08 #560 @trangdata
openalexR: an R package to interface with the OpenAlex API
Submitting Author Name: Trang Le
Submitting Author Github Handle: @trangdata
Repository:  https://github.com/massimoaria/openalexR
Version submitted: 1.0.2
Submission type: Standard
Due date for
Version accepted: TBD
---

```
Package: openalexR
Title: Getting Bibliographic Records from 'OpenAlex' Database Using 'DSL'
    API
Version: 1.0.2
    person(given = "Massimo",
           family = "Aria",
           email = "aria@unina.it",
    person(given = "Corrado",
           family = "Cuccurullo",
           role = c("ctb"),
           email = "cuccurullocorrado@gmail.com",
    person(given = "Trang",
           family = "Le",
           role = "aut",
           email = "grixor@gmail.com",
    )
Description: A set of tools to extract bibliographic content from
    'OpenAlex' database using API <https://docs.openalex.org/api/>.
    https://massimoaria.github.io/openalexR/
    curl,
    httr,
    progress,
    tibble
    covr
    knitr
```

- Explain how and why the package falls under these categories (briefly, 1-2 sentences): The package interacts with the [OpenAlex API](https://docs.openalex.org/api). Similar packages in the same category are rcrossref and rotl.

Anyone who wants to work in R to interact with the OpenAlex API to acquire information on publications, authors, etc., including researchers in the field of bibliometrics, text mining, etc. We include several example analyses in our README.

Please see the detailed answer [here](https://github.com/ropensci/software-review/issues/557#issuecomment-1295068888) with a detailed table of comparison and examples of code differences.

Yes.

https://github.com/ropensci/software-review/issues/557

N/A. [Passed all checks](https://github.com/massimoaria/openalexR/issues/43).
