### 2022-11-08 #560 @annakrystalli
Thanks for submitting @trangdata! 🎉

I'm in the process of assigning a handling editor.

In the meantime, it appears there's a minor discrepancy in the version stated in the submission and pasted `DESCRIPTION` (1.0.1) and the actual current package version (1.0.2). Would you mind correcting and double-checking the pasted `DESCRIPTION` is up to date?
