### 2022-11-08 #560 @trangdata
Thanks so much @annakrystalli! 🌻

> it appears there's a minor discrepancy in the version stated in the submission and pasted DESCRIPTION (1.0.1) and the actual current package version (1.0.2).

Ah my bad. I forgot to update the version. Edited now.
