### 2022-11-22 3585ba4
na.omit to fix #51
