+++ R/oa_snowball.R
@@ -94,6 +94,8 @@ oa_snowball <- function(identifier = NULL,
   edges <- rbind(citing_rel, cited_rel)
   # remove duplicates when two input identifiers cite each other
   edges <- edges[!duplicated(edges), ]
+  # remove edges to/from NA nodes
+  edges <- na.omit(edges)
 
   if (id_type == "short") {
     edges$to <- shorten_oaid(edges$to)
