### 2022-11-22 eb6f331
use complete.cases to avoid unnecessary attrs
