+++ man/ngram2df.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/oa_ngram.R
+\name{ngram2df}
+\alias{ngram2df}
+\title{Convert a "Work" entity's ngram data from list to data frame}
+\usage{
+ngram2df(ngram)
+}
+\description{
+Convert a "Work" entity's ngram data from list to data frame
+}
+\keyword{internal}
