+++ NAMESPACE
@@ -4,6 +4,7 @@ export(oa2bibliometrix)
 export(oa2df)
 export(oa_entities)
 export(oa_fetch)
+export(oa_ngrams)
 export(oa_query)
 export(oa_random)
 export(oa_request)
