+++ DESCRIPTION
@@ -45,7 +45,7 @@ VignetteBuilder:
     knitr
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.2.3
+RoxygenNote: 7.2.1
 Config/testthat/edition: 3
 Depends: 
     R (>= 2.10)
