### 2023-01-06 c8e4ae1
read json directly for speed
