+++ man/ngram2df.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/oa_ngram.R
+% Please edit documentation in R/oa_ngrams.R
 \name{ngram2df}
 \alias{ngram2df}
 \title{Convert a "Work" entity's ngram data from list to data frame}
