+++ man/oa_ngrams.Rd
@@ -1,17 +1,16 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/oa_ngram.R
+% Please edit documentation in R/oa_ngrams.R
 \name{oa_ngrams}
 \alias{oa_ngrams}
 \title{Get N-grams of works}
 \usage{
-oa_ngrams(works_identifier, mailto = oa_email(), verbose = FALSE)
+oa_ngrams(works_identifier, ..., verbose = FALSE)
 }
 \arguments{
 \item{works_identifier}{Character. OpenAlex ID(s) of "works" entities as item identifier(s).
 These IDs normally start with "W". See more at <https://docs.openalex.org/about-the-data/work#id>.}
 
-\item{mailto}{Character string.
-Gives OpenAlex an email to enter the polite pool.}
+\item{...}{Unused.}
 
 \item{verbose}{Logical.
 If TRUE, print information about the querying process. Defaults to TRUE.}
