+++ README.Rmd
@@ -346,7 +346,7 @@ ggraph(graph = as_tbl_graph(snowball_docs), layout = "stress") +
 
 ## N-grams
 
-OpenAlex offers (limited) support for full-text N-grams of Work entities (these have IDs starting with `"W"`). Given a vector of work IDs, `oa_ngrams()` returns a dataframe of N-gram data (in the `ngrams` list-column) for each work:
+OpenAlex offers (limited) support for [fulltext N-grams](https://docs.openalex.org/api-entities/works/get-n-grams#fulltext-coverage) of Work entities (these have IDs starting with `"W"`). Given a vector of work IDs, `oa_ngrams` returns a dataframe of N-gram data (in the `ngrams` list-column) for each work.
 
 ```{r}
 ngrams_data <- oa_ngrams(
@@ -356,7 +356,7 @@ ngrams_data <- oa_ngrams(
 
 ngrams_data
 
-lapply(ngrams$ngrams, head, 3)
+lapply(ngrams_data$ngrams, head, 3)
 
 ngrams_data |>
   unnest(ngrams) |>
@@ -373,6 +373,8 @@ ngrams_data |>
     y = NULL
   )
 ```
+
+`oa_ngrams` can sometimes be slow because the N-grams data can get pretty big, but given that the N-grams are `"cached via CDN"`](https://docs.openalex.org/api-entities/works/get-n-grams#api-endpoint), you may also consider parallelizing for this special case.
 
 
 ## About OpenAlex
