### 2023-01-14 e776b33
add in the clean-up line that got left out
