### 2023-01-17 b3541e4
issue #65: solved
in Mac and Linux using the original work_identifiers instead of normalized_id creates an error in file temp paths.
