### 2023-01-20 #560 @pachadotdev
> Hi @pachadotdev and @BriannaLind , checking in to see if you could update us on an approximate timeline for when you'll be able to do the review.

hi
i finished the wilsonfreitas review last weekend
i can check this one on monday
