### 2023-01-31 #560 @trangdata
Hi @BriannaLind, first of all, thank you so much for spending your time reviewing the package. ❤️  ✨ I appreciate your comments!

> The first example executed fine, but instead of the nice display shown in the tutorial I had a bunch of (kind of ugly tabsets).

Ah this is interesting. I'm not sure what's going on with knitr, but you're right: we should separate oa_fetch() and the lines to display the output dataframe, so that the user can still see the output any other ways. I will update the readme.

> The rest of the tutorial worked well, except several times I had to install packages that were not shown in the script but were necessary (this included ggraph, ggtext, and word cloud)

We had a few discussions regarding finding a balance between adding dependencies and making the user separately install things. We decided to keep the package minimal with few dependencies since most use cases won't require additional packages. When the user wants more specific analysis, e.g., snowballing, they would install ggraph separately. The examples were merely to show the variety of use cases for openalexR, so we didn't think attaching, for example, wordcloud, as an `Imports` was necessary. However, if you think these examples should be moved to a different place than readme, we will consider the suggestion!

Thanks again for the constructive feedback!
