### 2023-01-31 #560 @trangdata
> I think its totally fine for users to install extra packages separately - I mostly was thinking of a 'new R user' who is going to want to really work through your examples and copy and paste the script ---that it would be nice to have library(XXX) above the code snippet to reduce frustration.

@BriannaLind ah gotcha! That makes sense. I'll keep this in mind when updating the readme.

>  ll try to have my review done by Thu

@pachadotdev Thank you!!! I'm excited to hear your feedback! 🙌🏽
