### 2023-01-31 massimoaria/openalexR#67 @trangdata
Update readme to separate core and utility functions
Reviewer's comment [here](https://github.com/ropensci/software-review/issues/560#issuecomment-1409244254).
