### 2023-01-31 746f456
update readme - render Rmd to include ngram - separate core and show_w* functions
