+++ README.md
@@ -90,21 +90,39 @@ filter](https://massimoaria.github.io/openalexR/articles/Filters.html#works)
 (either the canonical form with <https://doi.org/> or without):
 
 ``` r
-oa_fetch(
-  doi = c("10.1016/j.joi.2017.08.007", "https://doi.org/10.1093/bioinformatics/btab727"),
+works_from_dois <- oa_fetch(
+  doi = c("10.1016/j.joi.2017.08.007", "https://doi.org/10.1007/s11192-013-1221-3"),
   entity = "works",
   verbose = TRUE
-) %>%
-  show_works() %>%
-  knitr::kable()
-#> Requesting url: https://api.openalex.org/works?filter=doi%3A10.1016%2Fj.joi.2017.08.007%7Chttps%3A%2F%2Fdoi.org%2F10.1093%2Fbioinformatics%2Fbtab727
+)
+#> Requesting url: https://api.openalex.org/works?filter=doi%3A10.1016%2Fj.joi.2017.08.007%7Chttps%3A%2F%2Fdoi.org%2F10.1007%2Fs11192-013-1221-3
 #> Getting 1 page of results with a total of 2 records...
 ```
 
-| id          | display_name                                                                           | first_author     | last_author        | so                      | url                                              | is_oa | top_concepts                                                       |
-|:------------|:---------------------------------------------------------------------------------------|:-----------------|:-------------------|:------------------------|:-------------------------------------------------|:------|:-------------------------------------------------------------------|
-| W2755950973 | bibliometrix : An R-tool for comprehensive science mapping analysis                    | Massimo Aria     | Corrado Cuccurullo | Journal of Informetrics | <https://doi.org/10.1016/j.joi.2017.08.007>      | FALSE | Data science                                                       |
-| W3206431085 | PMLB v1.0: an open-source dataset collection for benchmarking machine learning methods | Joseph D. Romano | Jason H. Moore     | Bioinformatics          | <https://doi.org/10.1093/bioinformatics/btab727> | TRUE  | Benchmarking, Python (programming language), Benchmark (surveying) |
+We can view the output tibble/dataframe, `works_from_dois`,
+interactively in RStudio or inspect it with base functions like `str` or
+`head`. We also provide the experimental `show_works` function to
+simplify the result (e.g., remove some columns, keep first/last author)
+for easy viewing.
+
+*Note*: the following table is wrapped in `knitr::kable()` to be
+displayed nicely in this README, but you will most likely not need this
+function.
+
+``` r
+# str(works_from_dois, max.level = 2)
+# head(works_from_dois)
+# show_works(works_from_dois) 
+
+works_from_dois |>
+  show_works() |>
+  knitr::kable()
+```
+
+| id          | display_name                                                              | first_author      | last_author        | so                      | url | is_oa | top_concepts                          |
+|:------------|:--------------------------------------------------------------------------|:------------------|:-------------------|:------------------------|:----|:------|:--------------------------------------|
+| W2755950973 | bibliometrix : An R-tool for comprehensive science mapping analysis       | Massimo Aria      | Corrado Cuccurullo | Journal of Informetrics | NA  | FALSE | Data science                          |
+| W2038196424 | Coverage and adoption of altmetrics sources in the bibliometric community | Stefanie Haustein | Jens Terliesner    | Scientometrics          | NA  | FALSE | Altmetrics, Bookmarking, Social media |
 
 **Goal**: Download all works published by a set of authors (known
 ORCIDs).
[…]
