### 2023-02-03 98cf27f
put check back on windows-latest
