### 2023-02-10 #560 @ldecicco-USGS
Thanks @pachadotdev and @BriannaLind for the reviews!

It seems like you both gave the package a test drive and things are looking pretty good. I don't see any major suggestions that were put forth yet to fix. Is that right? When I first looked at the package via the editor review, I also thought it was in pretty good shape!

If that's correct, @pachadotdev and @BriannaLind , could you use the Approval template and verify:
https://devguide.ropensci.org/approval2template.html
