+++ tests/testthat/test-oa_fetch.R
@@ -47,6 +47,8 @@ test_that("oa_fetch works", {
     paste0("https://openalex.org/", sort(work_ids))
   )
 
+  Sys.sleep(1/10)
+
   expect_s3_class(
     oa_fetch(
       entity = "w",
@@ -68,6 +70,7 @@ test_that("Error when input entity can't be matched", {
     )
   )
 
+  Sys.sleep(1/10)
   expect_error(
     oa_fetch(
       entity = "insta",
@@ -91,7 +94,7 @@ test_that("oa_fetch authors can deal with NA institutions", {
     ),
     "data.frame"
   )
-
+  Sys.sleep(1/10)
   expect_type(
     oa_fetch(
       entity = "authors",
@@ -158,7 +161,7 @@ test_that("oa_fetch can combine (OR) more than 50 DOIs in a filter", {
     "https://doi.org/10.1016/j.net.2020.08.005",
     "https://doi.org/10.1016/j.iot.2020.100318"
   )
-
+  Sys.sleep(1/10)
   many_doi_results <- oa_fetch(entity = "works", doi = valid_dois)
 
   expect_s3_class(
