### 2023-02-12 03f7f7f
simplify endpoint
