+++ man/oa2bibliometrix.Rd
@@ -36,7 +36,7 @@ query <- oa_query(
   from_publication_date = "2021-01-01",
   to_publication_date = "2021-12-31",
   search = NULL,
-  endpoint = "https://api.openalex.org/"
+  endpoint = "https://api.openalex.org"
 )
 
 res <- oa_request(
