+++ R/oa_fetch.R
@@ -55,8 +55,8 @@ oa_entities <- function() {
 #'   verbose = TRUE
 #' )
 #' }
-oa_fetch <- function(identifier = NULL, ## identifier of a work, author, venue, etc.
-                     entity = if (is.null(identifier)) NULL else id_type(identifier[[1]]),
+oa_fetch <- function(entity = if (is.null(identifier)) NULL else id_type(identifier[[1]]),
+                     identifier = NULL,
                      ...,
                      search = NULL,
                      sort = NULL,
@@ -461,7 +461,7 @@ oa_request <- function(query_url,
 
 oa_query <- function(filter = NULL,
                      multiple_id = FALSE,
-                     identifier = NULL, ## identifier of a work, author, venue, etc.
+                     identifier = NULL,
                      entity = if (is.null(identifier)) NULL else id_type(identifier[[1]]),
                      search = NULL,
                      sort = NULL,
