+++ tests/UArizona-1.R
@@ -1,0 +1,289 @@
+##################################################
+######## Author: Yan Han 
+######## Date: Feb 22, 2023
+
+# OpenAlex R
+# OpenAlex Documentation: https://docs.openalex.org/
+# Documentation: https://github.com/massimoaria/openalexR/tree/b3541e4f2695da771f15d0d92a7a050757e67e9c 
+
+
+# Install the developer version of openalexR from GitHub
+install.packages("remotes")
+remotes::install_github("massimoaria/openalexR")
+
+# Install the released version of openalexR from CRAN 
+install.packages("openalexR")
+
+# Before we go any further, we highly recommend you set openalexR.mailto option so that your requests go to the polite pool for faster response times: 
+# yhan: No documentation link to the polite pool. So I am not sure if an email will get you a higher priority to get response??? 
+options (openalexR.mailto="yhan@arizona.edu")
+
+# common libraries to add
+library(openalexR)
+library(dplyr)
+library(ggplot2)
+
+############################# DOI as filter ########################################
+# Getting record using DOIs
+works_from_dois <- oa_fetch( 
+  doi = c("https://doi.org/10.6017/ital.v40i1.12553", "10.1016/j.joi.2017.08.007"),
+  entity = "works",
+  verbose = TRUE
+  ) 
+
+# show the records
+str(works_from_dois, max.level=2) 
+head(works_from_dois)
+show_works(works_from_dois)
+
+works_from_dois |> 
+  show_works() |>
+  knitr::kable()
+
+#> Requesting url: https://api.openalex.org/works?filter=doi%3A10.1016%2Fj.joi.2017.08.007%7Chttps%3A%2F%2Fdoi.org%2F10.1093%2Fbioinformatics%2Fbtab727
+
+
+############################## ORCID as filter ####################################
+# Download all works published by a set of authors using ORCIDs
+# use author.orcid as a filter
+works_from_dois <- oa_fetch(
+  entity = "works",
+  #author.orcid = c("0000-0001-9518-2684"),
+  author.orcid = c("0000-0001-9518-2684", "0000-0002-8517-9411"),
+  verbose = TRUE
+) 
+
+works_from_dois |>
+  show_works()  |>
+  knitr::kable()
+#> Requesting url: https://api.openalex.org/works?filter=author.orcid%3A0000-0001-9518-2684%7C0000-0002-8517-9411
+# > output to a file 
+
+
+############################ Authors ################################################33
+###### Getting all works published by a set of authors using ORCIDs
+works_from_orcids <- oa_fetch(
+  entity = "works",
+  author.orcid = c("0000-0001-9518-2684"), # Only 1 record. not sure why? 
+  #author.orcid = c("0000-0001-9518-2684", "0000-0003-1613-5981"),
+  verbose = TRUE
+)  
+
+#> Requesting url: https://api.openalex.org/works?filter=author.orcid%3A0000-0001-6187-6610%7C0000-0002-8517-9411
+#> Getting 2 pages of results with a total of 211 records...
+
+works_from_orcids |> 
+  show_works() |> 
+  knitr::kable()
+
+##### Getting authors' info using their ORCIDs
+authors_from_orcids <- oa_fetch(
+  entity = "authors",
+  #orcid = c("0000-0001-9518-2684"),
+  #orcid = c("0000-0001-6187-6610", "0000-0002-8517-9411", "0000-0001-9518-2684") # Han record not right. not sure why
+  orcid =  c("0000-0001-6187-6610", "0000-0002-8517-9411", "0000-0003-1613-5981") # working right
+)
+
+authors_from_orcids |> 
+ show_authors() |>
+ knitr::kable()
+
[…]
