### 2023-02-23 #560 @ropensci-review-bot
Approved! Thanks @trangdata for submitting and @BriannaLind, @pachadotdev for your reviews! :grin:
