### 2023-02-23 #560 @ldecicco-USGS
@trangdata Congratulations on the great package. Sorry the timing went a little longer than usual. The "bot" automatically closes this issue when the package gets approved, but we can still use the issue to communicate (or feel free to re-open it if you have any questions).

Let me know if you have any questions on the instructions above.

Once again, thanks for submitting to rOpenSci and great work! 🎉
