### 2023-02-23 #560 @trangdata
@ldecicco-USGS Thank you so much! ❤️  And huge thank you to the reviewers @pachadotdev and @BriannaLind for your valuable feedback! 🌈

On the first TODO: @ldecicco-USGS Could you please add (or ask the bot to add) @massimoaria to ropensci so he could transfer "ownership" of the repository, please? The package currently lives at https://github.com/massimoaria/openalexR. Thank you!
