### 2023-02-24 #560 @massimoaria
Dear Maelle,
I just transferred the openalexR repository to the rOpenSci organization.

Thanks a lot
Massimo

⁣****************************************************
Dr. Massimo Aria
Full Professor in Statistics for Social Sciences
PhD in Computational Statistics
Bibliometrix R-package creator
K-Synth Academic Spin-Off co-founder
Department of Economics and Statistics
University of Naples Federico II
Monte S. Angelo, via Cinthia I-80126 Napoli, Italy
Room D-25, Sector D, 2nd Floor, Building 3
ph.  +39 081675187
fax  +39 081675009
mob. +39 392 1966384
email ***@***.***
https://www.massimoaria.com
https://www.bibliometrix.org
https://www.k-synth.unina.it
https://scholar.google.it/citations?user=Qu66YZQAAAAJ
https://orcid.org/0000-0002-8517-9411
http://www.researcherid.com/rid/O-7983-2015
https://www.researchgate.net/profile/Massimo_Aria
****************************************************​

> Il giorno 24 feb 2023, alle ore 06:30, Maëlle Salmon ***@***.***> ha scritto:
>
>
> I sent an invitation to the organization to @massimoaria <https://github.com/massimoaria>. Please ping me when the repo is transferred so that I might give @massimoaria <https://github.com/massimoaria> admin access back.
>
> Sorry for the clunkiness, it's only the second time in a short time we encounter this case ropensci-org/buffy#98 <https://github.com/ropensci-org/buffy/issues/98>
